requirements-dev.txt
# Debugging
debugpy

# Dependency maintenance
pip-tools

# Testing
pytest
pytest-cov
pytest-qgis

# Linting and formatting
pre-commit
mypy
ruff
flake8
flake8-qgis

# Stubs
PyQt5-stubs
pandas-stubs

# For running within qgis
# geopandas
# networkx

qgis-plugin-ci

sphinx
pydata-sphinx-theme
myst-parser
livereload
sphinx-toolbox
enum-tools
